requirements/requirements-django.txt
uWSGI
Django
django-compressor
django-versatileimagefield
crispy-bootstrap5
django-debug-toolbar
djangorestframework
markdown
django-filter
